[클라이언트] 구조 변경3
- 이벤트 구조 변경
  - 이벤트 함수에 파라미터 추가
  - ex. OnSceneChange 이벤트 함수는 변경 후 씬의 포인터를 파라미터로 가짐
- ButtonID, CharacterID, GameObjectID 자료형 추가
- 일부 FLOAT -> float로 변경
- 빌드 경고 문구 나오는 부분 수정

+++ Client/EditCtrl.cpp
@@ -198,7 +198,7 @@ void EditCtrl::MoveCaret(int distance)
 void EditCtrl::CreateTextLayout()
 {
 	auto dwriteFactory{ GameApp::GetInstance()->GetDwriteFactory() };
-	dwriteFactory->CreateTextLayout(m_text.c_str(), static_cast<UINT32>(m_text.length()), m_font->GetTextFormat().Get(), m_size.x, m_size.y, &m_textLayout);
+	dwriteFactory->CreateTextLayout(m_text.c_str(), static_cast<UINT32>(m_text.length()), m_font->GetTextFormat().Get(), static_cast<float>(m_size.x), static_cast<float>(m_size.y), &m_textLayout);
 	m_textLayout->SetParagraphAlignment(DWRITE_PARAGRAPH_ALIGNMENT_CENTER);
 	m_textLayout->SetWordWrapping(DWRITE_WORD_WRAPPING_NO_WRAP);
 }
